+++ source/os/memoryManager.ts
@@ -6,8 +6,6 @@ module DOGES {
 
     // Loads the program into memory
     public static loadProgram(programInput, priority): void {
-      // Return a memory bound violation if Resident List is full
-
       // Create a PCB
       var newPcb = new Pcb();
 
@@ -15,11 +13,11 @@ module DOGES {
       newPcb.base = this.fetchFreeBlock();
       newPcb.priority = priority;
 
-      if (newPcb.base === null) {
-        // _StdOut.putText("Memory very full. Cannot load. Much sadness.");
+      if (newPcb.base === -1) {
         var processFilename = ProcessManager.createProcessFilename(newPcb);
         _FileSystem.createFile(processFilename);
         newPcb.inFileSystem = true;
+        newPcb.limit = -1;
 
         _FileSystem.writeFile(processFilename, programInput);
       } else {
@@ -37,8 +35,6 @@ module DOGES {
     }
 
     public static loadToMemory(programInput, startPoint): void {
-      programInput = programInput.replace(/\s/g, "").toUpperCase();
-
       for (var i = 0; i < programInput.length / 2; i++) {
         if (i === 0) {
           var currentCode: number = 0;
@@ -72,41 +68,76 @@ module DOGES {
 
     // Takes the program out of the file system and puts it in main memory
     // for continued execution
-    public static rollIn(program): void {
+    // Returns a boolean of the roll in status
+    public static rollIn(program): boolean {
+        _Kernel.krnTrace("Rolling in process " + program.PID + ".");
+
         program.base = this.fetchFreeBlock();
+
+        // No free memory available
+        if (program.base === null || program.base === -1) {
+            return false;
+        }
 
         // Look up the program from the file system
         var programFilename = ProcessManager.createProcessFilename(program);
-        var data = _FileSystem.readFile(programFilename);
+        var readFile = _FileSystem.readFile(programFilename);
+
+        // Failed to read data from file
+        if (readFile.status === "ERROR") {
+            return false;
+        }
 
         // Then load the program from the file system to main memory
-        this.loadToMemory(data, program.base);
-        _FileSystem.deleteFile(programFilename);
+        this.loadToMemory(readFile.body, program.base);
+        var deleteFile = _FileSystem.deleteFile(programFilename);
+
+        // File can't be deleted in file system
+        if (deleteFile.status === "ERROR") {
+            return false;
+        }
 
         program.limit = program.base + PROGRAM_SIZE - 1;
         program.inFileSystem = false;
+
+        return true;
     }
 
     // Takes the program out of main memory and stores it in the file system
     // for later execution
-    public static rollOut(program): void {
+    // Returns a boolean of the roll out status
+    public static rollOut(program): boolean {
+        _Kernel.krnTrace("Rolling out process " + program.PID + ".");
+
         // Create a new file for the rolled-out program
         var programFilename = ProcessManager.createProcessFilename(program);
         var programInput = this.fetchProgramInput(program.base);
-        _FileSystem.createFile(programFilename);
+        var createdFile = _FileSystem.createFile(programFilename);
+
+        // File can't be created in file system
+        if (createdFile.status === "ERROR") {
+            return false;
+        }
 
         // Then write the program data into the file system
-        _FileSystem.writeFile(programFilename, programInput);
+        var writtenFile = _FileSystem.writeFile(programFilename, programInput);
+
+        // File can't be written in file system
+        if (writtenFile.status === "ERROR") {
+            return false;
+        }
 
         program.state = PS_READY;
-        program.base = null;
-        program.limit = null;
+        program.base = -1;
+        program.limit = -1;
         program.inFileSystem = true;
+
+        return true;
     }
 
     // Returns the program input in that particular memory segment
     public static fetchProgramInput(startAddress): string {
-        var programInput: string = "";
+        var programInput = "";
 
         for (var i = startAddress; i < startAddress + PROGRAM_SIZE; i++) {
             programInput += _Memory.memArray[i];
@@ -122,11 +153,11 @@ module DOGES {
         for (var i = 0; i < _ResidentList.length; i++) {
           if ( _ResidentList[i] !== undefined
             && _ResidentList[i].base === freeAddress
-            && _ResidentList[i].base !== null) {
+            && _ResidentList[i].base !== -1) {
             freeAddress += PROGRAM_SIZE;
 
             if (freeAddress >= PROGRAM_SIZE * PROGRAM_LIMIT) {
-                freeAddress = null;
+                freeAddress = -1;
             }
           }
         }
@@ -142,7 +173,7 @@ module DOGES {
 
     // Clears one memory segment in main memory
     public static clearSegment(startPoint): void {
-        if (startPoint === null || startPoint === undefined) {
+        if (startPoint === null || startPoint === undefined || startPoint === -1) {
             startPoint = 0;
         }
